rename GameLayout to GameBody
