+++ src/components/game/GameBody/ContainerTabs.tsx
@@ -1,6 +1,6 @@
 import type { PlayerFull } from 'types/game.d'
 import { GameCard, TabLabels } from '@prisma/client'
-import { TabsWrapper, TabStyle } from './GameLayoutStyles'
+import { TabsWrapper, TabStyle } from './GameBodyStyles'
 import { getBoard } from '../shared/game.utils'
 import { cardCounter } from 'assets/strings'
