+++ src/components/game/index/Game.tsx
@@ -2,7 +2,7 @@ import type { Game } from '@prisma/client'
 import type { ServerProps, BasicLands, PartialGame } from 'types/game'
 import GameHeader from 'components/game/GameHeader/GameHeader'
 import PlayerJoin from 'components/game/PlayerJoin/PlayerJoin'
-import GameLayout from 'components/game/GameBody/GameLayout'
+import GameLayout from 'components/game/GameBody/GameBody'
 import LandsModal from 'components/game/LandsModal/LandsModal'
 import HostModal from 'components/game/HostModal/HostModal'
 import LogModal from '../LogModal/LogModal'
